Corrección de comparación de primitivias

+++ src/Umeca/src/main/java/com/umeca/service/shared/UpDwFileGenericServiceImpl.java
@@ -119,7 +119,7 @@ public class UpDwFileGenericServiceImpl implements UpDwFileGenericService {
         extension = extension.toLowerCase();
         final Long fileTypeId = catFileTypeRepository.findByExtension(extension);
 
-        if (fileTypeId == null || fileTypeId <= 0) {
+        if (fileTypeId == null || fileTypeId.longValue() <= 0) {
             resMsg.setMessage("Tipo de archivo no permitido");
             resMsg.setHasError(true);
             return false;
@@ -145,7 +145,7 @@ public class UpDwFileGenericServiceImpl implements UpDwFileGenericService {
         extension = extension.toLowerCase();
         final Long fileTypeId = catFileTypeRepository.findByExtensionCode(extension, code);
 
-        if (fileTypeId == null || fileTypeId <= 0) {
+        if (fileTypeId == null || fileTypeId.longValue() <= 0) {
             resMsg.setMessage("Tipo de archivo no permitido");
             resMsg.setHasError(true);
             return false;
